Add `FileSpec` type
In `globTestSpecs`, the filters will be matched against the files, and
if there is a location in the filter, it will be included with its file,
in a `FileSpec: { path, testLocations? }`.

I thought it would be a good idea to include the files with
the location in the same object, instead of having them in their own
attribute in the config and then matching based on the path or
something.  But I ended up having to modify types everywhere to get this
to work.

I also had to modify `groupFilesByEnv` to include the `testLocations` in
its output, and also modify pools to pass the test locations to
`runFiles`.

Much of the code in this commit is hacked together, but I had to get
something that somewhat works, sometimes, just to see if I'm on the
right track.